Make position selector keep state across alliance switch changes
The position chooser now uses an enum of "1", "2", and "3" to keep track of which position is selected, changing only the labels when the alliance changes. This way, an option will remain selected when the alliance is changed, instead of previous behavior of having no radio button selected after the alliance switch is used.
Additionally, this commit updates some styles for general appearance improvements.